Storage: move dir_walk to utils folder
