+++ applications/services/storage/storage_external_api.c
@@ -2,9 +2,9 @@
 #include <core/record.h>
 #include <storage/storage.h>
 #include <storage/storage_i.h> // IWYU pragma: keep
+#include <storage/util/dir_walk.h>
 #include <storage/storage_message.h>
 #include <toolbox/stream/file_stream.h>
-#include <toolbox/dir_walk.h>
 #include <toolbox/path.h>
 
 #define MAX_NAME_LENGTH  256
