+++ applications/services/storage/util/dir_walk.c
