+++ lib/toolbox/SConscript
@@ -16,6 +16,7 @@ env.Append(
         File("float_tools.h"),
         File("hex.h"),
         File("md5_calc.h"),
+        File("path.h"),
         File("property.h"),
         File("rle_encode.h"),
         File("sha256_calc.h"),
@@ -36,6 +37,7 @@ sources = [
     "float_tools.c",
     "hex.c",
     "md5_calc.c",
+    "path.c",
     "property.c",
     "rle_encode.c",
     "sha256_calc.c",
